# public/openapi.yaml
openapi: 3.0.0
info:
  title: ZulaGap-AI API
  version: 1.1.0
paths:
  /api/teams:
    get:
      summary: 팀 목록 조회
      responses:
        '200':
          description: 팀 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Team'
        '400':
          description: 잘못된 요청
        '500':
          description: 서버 오류
    post:
      summary: 팀 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TeamCreate'
      responses:
        '200':
          description: 생성된 팀
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Team'
        '400':
          description: 잘못된 요청
        '500':
          description: 서버 오류
...
  /api/brands:
    get:
      summary: 브랜드 목록 조회
      responses:
        '200':
          description: 브랜드 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Brand'
    post:
      summary: 브랜드 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BrandCreate'
      responses:
        '200':
          description: 생성된 브랜드
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Brand'
...
  /api/agents:
    get:
      summary: 에이전트 목록 조회
      responses:
        '200':
          description: 에이전트 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Agent'
    post:
      summary: 에이전트 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentCreate'
      responses:
        '200':
          description: 생성된 에이전트
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
...
  /api/workflows:
    get:
      summary: 워크플로우 목록 조회
      responses:
        '200':
          description: 워크플로우 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Workflow'
    post:
      summary: 워크플로우 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WorkflowCreate'
      responses:
        '200':
          description: 생성된 워크플로우
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workflow'
...
  /api/documents:
    get:
      summary: 문서 목록 조회
      responses:
        '200':
          description: 문서 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Document'
    post:
      summary: 문서 생성
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentCreate'
      responses:
        '200':
          description: 생성된 문서
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Document'
...
components:
  schemas:
    Team:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        brandId: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    TeamCreate:
      type: object
      required: [name, brandId]
      properties:
        name: { type: string }
        description: { type: string }
        brandId: { type: string }
    Brand:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        identity: { type: object }
    BrandCreate:
      type: object
      required: [name]
      properties:
        name: { type: string }
        identity: { type: object }
...
    Agent:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        userId: { type: string }
        teamId: { type: string }
        config: { type: object }
        ragDocs:
          type: array
          items: { type: string }
        prompt: { type: string }
        type: { type: string }
        role: { type: string }
        parentAgentId: { type: string }
        identity: { type: object }
        isPublic: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    AgentCreate:
      type: object
      required: [name, userId, teamId, type, role]
      properties:
        name: { type: string }
        description: { type: string }
        userId: { type: string }
        teamId: { type: string }
        config: { type: object }
        ragDocs:
          type: array
          items: { type: string }
        prompt: { type: string }
        type: { type: string }
        role: { type: string }
        parentAgentId: { type: string }
        identity: { type: object }
    Workflow:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string }
        userId: { type: string }
        teamId: { type: string }
        teamLeaderType: { type: string }
        leaderAgentId: { type: string }
        config: { type: object }
    WorkflowCreate:
      type: object
      required: [name, userId, teamId, teamLeaderType]
      properties:
        name: { type: string }
        description: { type: string }
        userId: { type: string }
        teamId: { type: string }
        teamLeaderType: { type: string }
        leaderAgentId: { type: string }
        config: { type: object }
...
    Document:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        content: { type: string }
        source: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
        userId: { type: string }
        teamId: { type: string }
        brandId: { type: string }
        metadata: { type: object }
        relations: { type: object }
        embedding: { type: object }
        agentId: { type: string }
    DocumentCreate:
      type: object
      required: [title, content]
      properties:
        title: { type: string }
        content: { type: string }
        source: { type: string }
        userId: { type: string }
        teamId: { type: string }
        brandId: { type: string }
        metadata: { type: object }
        relations: { type: object }
        embedding: { type: object }
        agentId: { type: string }
...